get the rest of the timestamps for player status
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18816 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/Player.h
@@ -183,7 +183,7 @@ public:
   void*		pack(void*, uint16_t& code);
   void*		unpack(void*, uint16_t code);
 
-  void		setDeadReckoning(float timestamp);
+  void		setDeadReckoning(double timestamp);
 
   void		setUserTexture ( const char *tex ) { if(tex) userTexture = tex;}
